refactor: #6 BusRouteLineView 컴포넌트화
1. 추후 알람화면의 테이블뷰 셀에서 재사용 가능성이 있는 버스 라인 뷰를 컴포넌트화 하였음.
