+++ BBus/BBus/BusRoute/View/BusStationTableViewCell.swift
@@ -12,14 +12,7 @@ class BusStationTableViewCell: UITableViewCell {
     static let reusableID = "BusStationTableViewCell"
     static let cellHeight: CGFloat = 80
 
-    private lazy var beforeLineView = UIView()
-    private lazy var afterLineView = UIView()
-
-    private lazy var centerPointImageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.image = UIImage(named: "StationCenterCircle")
-        return imageView
-    }()
+    private lazy var busRouteLineView = BusRouteLineView()
     
     private lazy var stationTitleLabel: UILabel = {
         let label = UILabel()
@@ -48,39 +41,20 @@ class BusStationTableViewCell: UITableViewCell {
         self.selectionStyle = .none
     }
 
-    func configureLayout() {
-        self.addSubview(self.centerPointImageView)
-        self.centerPointImageView.translatesAutoresizingMaskIntoConstraints = false
+    private func configureLayout() {
+        self.addSubview(self.busRouteLineView)
+        self.busRouteLineView.translatesAutoresizingMaskIntoConstraints = false
         NSLayoutConstraint.activate([
-            self.centerPointImageView.heightAnchor.constraint(equalToConstant: 15),
-            self.centerPointImageView.widthAnchor.constraint(equalTo: self.centerPointImageView.heightAnchor),
-            self.centerPointImageView.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 80),
-            self.centerPointImageView.centerYAnchor.constraint(equalTo: self.centerYAnchor)
-        ])
-        
-        self.insertSubview(self.beforeLineView, at: 0)
-        self.beforeLineView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            self.beforeLineView.heightAnchor.constraint(equalToConstant: BusStationTableViewCell.cellHeight/2),
-            self.beforeLineView.widthAnchor.constraint(equalToConstant: 3),
-            self.beforeLineView.topAnchor.constraint(equalTo: self.topAnchor),
-            self.beforeLineView.centerXAnchor.constraint(equalTo: self.centerPointImageView.centerXAnchor)
-        ])
-        
-        self.insertSubview(self.afterLineView, at: 0)
-        self.afterLineView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            self.afterLineView.heightAnchor.constraint(equalToConstant: BusStationTableViewCell.cellHeight/2),
-            self.afterLineView.widthAnchor.constraint(equalToConstant: 3),
-            self.afterLineView.bottomAnchor.constraint(equalTo: self.bottomAnchor),
-            self.afterLineView.leadingAnchor.constraint(equalTo: self.beforeLineView.leadingAnchor)
+            self.busRouteLineView.topAnchor.constraint(equalTo: self.topAnchor),
+            self.busRouteLineView.bottomAnchor.constraint(equalTo: self.bottomAnchor),
+            self.busRouteLineView.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 80)
         ])
         
         self.addSubview(self.stationTitleLabel)
         self.stationTitleLabel.translatesAutoresizingMaskIntoConstraints = false
         NSLayoutConstraint.activate([
             self.stationTitleLabel.topAnchor.constraint(equalTo: self.topAnchor, constant: 18),
-            self.stationTitleLabel.leadingAnchor.constraint(equalTo: self.centerPointImageView.trailingAnchor, constant: 15),
+            self.stationTitleLabel.leadingAnchor.constraint(equalTo: self.busRouteLineView.trailingAnchor, constant: 15),
             self.stationTitleLabel.trailingAnchor.constraint(equalTo: self.trailingAnchor, constant: -20)
         ])
         
@@ -101,8 +75,7 @@ class BusStationTableViewCell: UITableViewCell {
     }
     
     func configureLineColor(before: UIColor, after: UIColor) {
-        self.beforeLineView.backgroundColor = before
-        self.afterLineView.backgroundColor = after
+        self.busRouteLineView.configureLineColor(before: before, after: after)
     }
     
     func configureMockData() {
